+++ src/main/java/com/scottkillen/mod/dendrology/compat/chisel/ChiselMod.java
@@ -5,6 +5,7 @@ import com.cricketcraft.chisel.carving.Carving;
 import com.cricketcraft.chisel.init.ChiselTabs;
 import com.google.common.collect.ImmutableList;
 import com.scottkillen.mod.dendrology.TheMod;
+import com.scottkillen.mod.dendrology.config.Settings;
 import com.scottkillen.mod.dendrology.content.overworld.OverworldTreeSpecies;
 import com.scottkillen.mod.koresample.compat.Integrator;
 import cpw.mods.fml.common.Loader;
@@ -81,7 +82,7 @@ public final class ChiselMod extends Integrator
     @Override
     public void doIntegration(ModState modState)
     {
-        if (Loader.isModLoaded(MOD_ID))
+        if (Loader.isModLoaded(MOD_ID)&& Settings.INSTANCE.integrateChisel())
         {
             switch (modState)
             {
@@ -94,14 +95,8 @@ public final class ChiselMod extends Integrator
     }
 
     @Override
-    protected String modID()
-    {
-        return MOD_ID;
-    }
+    protected String modID() { return MOD_ID; }
 
     @Override
-    protected String modName()
-    {
-        return MOD_NAME;
-    }
+    protected String modName() { return MOD_NAME; }
 }