+++ src/main/java/com/scottkillen/mod/dendrology/compat/forestry/ForestryMod.java
@@ -1,6 +1,7 @@
 package com.scottkillen.mod.dendrology.compat.forestry;
 
 import com.scottkillen.mod.dendrology.TheMod;
+import com.scottkillen.mod.dendrology.config.Settings;
 import com.scottkillen.mod.dendrology.content.overworld.OverworldTreeSpecies;
 import com.scottkillen.mod.koresample.compat.Integrator;
 import cpw.mods.fml.common.Loader;
@@ -121,7 +122,7 @@ public final class ForestryMod extends Integrator
     @Override
     public void doIntegration(ModState modState)
     {
-        if (Loader.isModLoaded(MOD_ID))
+        if (Loader.isModLoaded(MOD_ID) && Settings.INSTANCE.integrateForestry())
         {
             switch (modState)
             {
@@ -137,14 +138,8 @@ public final class ForestryMod extends Integrator
     }
 
     @Override
-    protected String modID()
-    {
-        return MOD_ID;
-    }
+    protected String modID() { return MOD_ID; }
 
     @Override
-    protected String modName()
-    {
-        return MOD_NAME;
-    }
+    protected String modName() { return MOD_NAME; }
 }